Resolve comments
