+++ build.sh
@@ -17,6 +17,7 @@ else
 fi
 WITH_PYTHON_C2=${WITH_PYTHON_C2:=OFF}
 WITH_NNPACK=${WITH_NNPACK:=OFF}
+WITH_TAPIR=${WITH_TAPIR:=ON}
 PYTHON=${PYTHON:="`which python3`"}
 PROTOC=${PROTOC:="`which protoc`"}
 CORES=${CORES:=32}
@@ -401,6 +402,7 @@ function install_tc() {
     rm -rf *
     VERBOSE=${VERBOSE} ${CMAKE_VERSION} -DWITH_CAFFE2=${WITH_CAFFE2} \
         -DCMAKE_EXPORT_COMPILE_COMMANDS=ON \
+        -DWITH_TAPIR=${WITH_TAPIR} \
         -DPYTHON_EXECUTABLE=${PYTHON} \
         -DHALIDE_PREFIX=${INSTALL_PREFIX} \
         -DCMAKE_BUILD_TYPE=${BUILD_TYPE} \
